change names and folders
