+++ app/ideciclo/[id]/ideciclo_mapstyle.ts
