+++ app/ideciclo/[id]/ideciclo_radarstyle.js
