+++ app/ideciclo/[id]/ideciclo_rates.js
