+++ app/observatorio/calcs.js
