+++ public/dbs/ideciclo_rates_descriptions.json
