+++ servers.js
@@ -8,5 +8,6 @@ export const COUNTINGS_PAGE_DATA =`https://cms.ameciclo.org/contagens`
 export const IDECICLO_DATA = `https://api.ideciclo.ameciclo.org/reviews`
 export const IDECICLO_STRUCTURES_DATA = `https://api.ideciclo.ameciclo.org/structures`
 export const IDECICLO_PAGE_DATA = `https://cms.ameciclo.org/ideciclo`
+export const IDECICLO_FORMS_DATA = `https://api.ideciclo.ameciclo.org/forms`
 export const PERFIL_PAGE_DATA = `https://cms.ameciclo.org/perfil`
 export const PERFIL_DATA = `https://api.perfil.ameciclo.org/v1/cyclist-profile/summary/`
